+++ examples/values_dialog.py
@@ -1,0 +1,8 @@
+from magicgui.widgets import request_values
+
+vals = request_values(
+    age=int,
+    name=dict(annotation=str, label="Enter your name:"),
+    title="Hi, who are you?",
+)
+print(repr(vals))