+++ magicgui/widgets/__init__.py
@@ -18,6 +18,7 @@ from ._concrete import (
     Container,
     DateEdit,
     DateTimeEdit,
+    Dialog,
     EmptyWidget,
     FileEdit,
     FloatSlider,
@@ -42,7 +43,7 @@ from ._concrete import (
     TimeEdit,
     TupleEdit,
 )
-from ._dialogs import show_file_dialog
+from ._dialogs import request_values, show_file_dialog
 from ._function_gui import FunctionGui, MainFunctionGui
 from ._image import Image
 from ._table import Table
@@ -72,6 +73,7 @@ __all__ = [
     "create_widget",
     "DateEdit",
     "DateTimeEdit",
+    "Dialog",
     "EmptyWidget",
     "FileEdit",
     "FloatSlider",
@@ -101,6 +103,7 @@ __all__ = [
     "TupleEdit",
     "Widget",
     "show_file_dialog",
+    "request_values",
 ]
 
 del partial