+++ magicgui/widgets/_bases/container_widget.py
@@ -46,6 +46,9 @@ class ContainerWidget(Widget, _OrientationMixin, MutableSequence[Widget]):
     layout : str, optional
         The layout for the container.  must be one of ``{'horizontal',
         'vertical'}``. by default "vertical"
+    scrollable : bool, optional
+        Whether to enable scroll bars or not. If enabled, scroll bars will
+        only appear along the layout direction, not in both directions.
     widgets : Sequence[Widget], optional
         A sequence of widgets with which to intialize the container, by default
         ``None``.
@@ -62,6 +65,7 @@ class ContainerWidget(Widget, _OrientationMixin, MutableSequence[Widget]):
     def __init__(
         self,
         layout: str = "vertical",
+        scrollable: bool = False,
         widgets: Sequence[Widget] = (),
         labels=True,
         **kwargs,
@@ -69,7 +73,7 @@ class ContainerWidget(Widget, _OrientationMixin, MutableSequence[Widget]):
         self._list: list[Widget] = []
         self._labels = labels
         self._layout = layout
-        kwargs["backend_kwargs"] = {"layout": layout}
+        kwargs["backend_kwargs"] = {"layout": layout, "scrollable": scrollable}
         super().__init__(**kwargs)
         self.extend(widgets)
         self.parent_changed.connect(self.reset_choices)
@@ -281,6 +285,12 @@ class ContainerWidget(Widget, _OrientationMixin, MutableSequence[Widget]):
 
     NO_VALUE = "NO_VALUE"
 
+    def asdict(self) -> dict[str, Any]:
+        """Return state of widget as dict."""
+        return {
+            w.name: getattr(w, "value", None) for w in self if w.name and not w.gui_only
+        }
+
     @debounce
     def _dump(self, path):
         """Dump the state of the widget to `path`."""
@@ -332,3 +342,16 @@ class MainWindowWidget(ContainerWidget):
         ``menu_name`` will be created if it does not already exist.
         """
         self._widget._mgui_create_menu_item(menu_name, item_name, callback, shortcut)
+
+
+class DialogWidget(ContainerWidget):
+    """Modal Container."""
+
+    _widget: _protocols.DialogProtocol
+
+    def exec(self) -> bool:
+        """Show the dialog, and block.
+
+        Return True if the dialog was accepted, False if rejected.
+        """
+        return bool(self._widget._mgui_exec())