+++ magicgui/widgets/_bases/ranged_widget.py
@@ -1,28 +1,37 @@
+import builtins
 from abc import ABC, abstractmethod
-from typing import Tuple, Union
+from math import ceil, log10
+from typing import Tuple, Union, cast
 from warnings import warn
 
 from magicgui.widgets import _protocols
 
-from .value_widget import UNSET, ValueWidget
+from .value_widget import UNSET, ValueWidget, _Unset
 
 
 class RangedWidget(ValueWidget):
-    """Widget with a contstrained value. Wraps RangedWidgetProtocol.
+    """Widget with a constrained value. Wraps RangedWidgetProtocol.
 
     Parameters
     ----------
     min : float, optional
-        The minimum allowable value, by default 0
+        The minimum allowable value, by default 0 (or `value` if `value` is less than 0)
     max : float, optional
-        The maximum allowable value, by default 1000
+        The maximum allowable value, by default 999 (or `value` if `value` is greater
+        than 999)
     step : float, optional
-        The step size for incrementing the value, by default 1
+        The step size for incrementing the value, by default adaptive step is used
     """
 
     _widget: _protocols.RangedWidgetProtocol
 
-    def __init__(self, min: float = 0, max: float = 1000, step: float = 1, **kwargs):
+    def __init__(
+        self,
+        min: Union[float, _Unset] = UNSET,
+        max: Union[float, _Unset] = UNSET,
+        step: Union[float, _Unset, None] = UNSET,
+        **kwargs,
+    ):  # sourcery skip: avoid-builtin-shadow
         for key in ("maximum", "minimum"):
             if key in kwargs:
                 warn(
@@ -38,9 +47,22 @@ class RangedWidget(ValueWidget):
         val = kwargs.pop("value", UNSET)
         super().__init__(**kwargs)
 
-        self.step = step
-        self.min = min
-        self.max = max
+        tmp_val = float(val if val not in (UNSET, None) else 1)
+
+        if step is UNSET or step is None:
+            self.step = None
+            self._widget._mgui_set_step(1)
+        else:
+            self.step = cast(float, step)
+
+        self.min: float = (
+            cast(float, min) if min is not UNSET else builtins.min(0, tmp_val)
+        )
+        self.max: float = (
+            cast(float, max)
+            if max is not UNSET
+            else builtins.max(1000, 10 ** ceil(log10(builtins.max(1, tmp_val + 1)))) - 1
+        )
         if val not in (UNSET, None):
             self.value = val
 
@@ -86,13 +108,31 @@ class RangedWidget(ValueWidget):
         self._widget._mgui_set_max(value)
 
     @property
-    def step(self) -> float:
-        """Step size for widget values."""
+    def step(self) -> Union[float, None]:
+        """Step size for widget values (None if adaptive step is turned on)."""
+        if self._widget._mgui_get_adaptive_step():
+            return None
         return self._widget._mgui_get_step()
 
     @step.setter
-    def step(self, value: float):
-        self._widget._mgui_set_step(value)
+    def step(self, value: Union[float, None]):
+        if value is None:
+            self._widget._mgui_set_adaptive_step(True)
+        else:
+            self._widget._mgui_set_adaptive_step(False)
+            self._widget._mgui_set_step(value)
+
+    @property
+    def adaptive_step(self):
+        """Whether the step size is adaptive."""
+        return self.step is None
+
+    @adaptive_step.setter
+    def adaptive_step(self, value: bool):
+        if value:
+            self.step = None
+        else:
+            self.step = self._widget._mgui_get_step()
 
     @property
     def range(self) -> Tuple[float, float]: